+++ packages/hooks/test/create-hook-app.test.js
@@ -1,13 +1,46 @@
 const { resetState } = require('../src/state');
-const { runHookApp } = require('../src/create-hook-app');
+const forrestjs = require('../src/index');
 const { registerAction } = require('../src/register-action');
+const {
+  isDeclarativeAction,
+  isListOfDeclarativeActions,
+} = require('../src/create-hook-app');
 const constants = require('../src/constants');
 
 describe('hooks/create-hook-app', () => {
   beforeEach(resetState);
 
+  describe('utils', () => {
+    const featureA = [
+      {
+        hook: 'foo',
+        handler: () => {},
+      },
+      {
+        hook: 'foo',
+        handler: [],
+      },
+      {
+        hook: 'foo',
+        handler: {},
+      },
+    ];
+
+    describe('isDeclarativeAction', () => {
+      featureA.forEach((payload) =>
+        it('Should recognise a declarative action payload', () => {
+          expect(isDeclarativeAction(payload)).toBe(true);
+        }),
+      );
+    });
+
+    describe('isListOfDeclarativeActions', () => {
+      expect(isListOfDeclarativeActions(featureA)).toBe(true);
+    });
+  });
+
   it('should run an empty app', async () => {
-    await runHookApp();
+    await forrestjs.run();
   });
 
   it('should register a service or feature from an ES module', async () => {
@@ -19,7 +52,7 @@ describe('hooks/create-hook-app', () => {
       },
     };
 
-    await runHookApp({
+    await forrestjs.run({
       services: [s1],
       features: [s1],
     });
@@ -31,7 +64,7 @@ describe('hooks/create-hook-app', () => {
     const handler = jest.fn();
     const s1 = () => handler();
 
-    await runHookApp({
+    await forrestjs.run({
       services: [s1],
       features: [s1],
     });
@@ -39,12 +72,13 @@ describe('hooks/create-hook-app', () => {
     expect(handler.mock.calls.length).toBe(2);
   });
 
-  it('should register a service as single hook setup', async () => {
+  // DEPRECATED
+  it.skip('should register a service as single hook setup', async () => {
     const handler = jest.fn();
     const s1 = ['foo', handler];
     const f1 = ({ createHook }) => createHook('foo');
 
-    await runHookApp({
+    await forrestjs.run({
       services: [s1],
       features: [f1],
     });
@@ -53,45 +87,54 @@ describe('hooks/create-hook-app', () => {
   });
 
   it('should run an app that provides settings as a function', async () => {
-    await runHookApp({
+    const f1 = jest.fn();
+    await forrestjs.run({
       settings: ({ setConfig }) => {
         setConfig('foo.faa', 22);
       },
       features: [
-        [
-          constants.START_FEATURE,
-          ({ getConfig }) => expect(getConfig('foo.faa')).toBe(22),
-        ],
+        {
+          hook: '$START_FEATURE',
+          handler: ({ getConfig }) => f1(getConfig('foo.faa')),
+        },
       ],
     });
+
+    expect(f1.mock.calls[0][0]).toBe(22);
   });
 
   it('should provide a config getter to any registered action', async () => {
-    await runHookApp({
+    const f1 = jest.fn();
+    await forrestjs.run({
       settings: ({ setConfig }) => {
         setConfig('foo.faa', 22);
       },
       services: [
-        // register a feature
+        // register a programmatic feature
         ({ registerAction }) =>
           registerAction({
-            hook: constants.INIT_SERVICE,
+            hook: '$INIT_SERVICE',
             handler: ({ getConfig, setConfig }) =>
               setConfig('foo', getConfig('foo.faa') * 2),
           }),
       ],
       features: [
-        // register a single action
-        [
-          constants.START_FEATURE,
-          ({ getConfig }) => expect(getConfig('foo')).toBe(44),
-        ],
+        // register a declarative feature
+        {
+          hook: '$START_FEATURE',
+          handler: ({ getConfig }) => f1(getConfig('foo')),
+        },
       ],
     });
+
+    expect(f1.mock.calls[0][0]).toBe(44);
   });
 
   it('should lock a context and decorate it with internal methods', async () => {
-    await runHookApp({
+    const f1 = jest.fn();
+    const f2 = jest.fn();
+    const f3 = jest.fn();
+    await forrestjs.run({
       settings: {
         increment: 1,
       },
@@ -99,26 +142,30 @@ describe('hooks/create-hook-app', () => {
         foo: (args, ctx) => args.value + ctx.getConfig('increment'),
       },
       services: [
-        [
-          constants.START_SERVICE,
-          async ({ createHook }) => {
+        {
+          hook: '$START_SERVICE',
+          handler: async ({ createHook }) => {
             const r1 = createHook.sync('aaa', { value: 1 });
-            expect(r1[0][0]).toBe(2);
+            f1(r1[0][0]);
 
-            const r2 = await createHook.serie('bbb', { value: 1 });
-            expect(r2[0][0]).toBe(2);
+            const r2 = await createHook.serie('bbb', { value: 2 });
+            f2(r2[0][0]);
 
-            const r3 = await createHook.parallel('ccc', { value: 1 });
-            expect(r3[0][0]).toBe(2);
+            const r3 = await createHook.parallel('ccc', { value: 3 });
+            f3(r3[0][0]);
           },
-        ],
+        },
       ],
       features: [
-        ['aaa', (args, ctx) => ctx.foo(args, ctx)],
-        ['bbb', (args, ctx) => ctx.foo(args, ctx)],
-        ['ccc', (args, ctx) => ctx.foo(args, ctx)],
+        { hook: 'aaa', handler: (args, ctx) => ctx.foo(args, ctx) },
+        { hook: 'bbb', handler: (args, ctx) => ctx.foo(args, ctx) },
+        { hook: 'ccc', handler: (args, ctx) => ctx.foo(args, ctx) },
       ],
     });
+
+    expect(f1.mock.calls[0][0]).toBe(2);
+    expect(f2.mock.calls[0][0]).toBe(3);
+    expect(f3.mock.calls[0][0]).toBe(4);
   });
 
   describe('createHookApp getters / setters', () => {
@@ -126,14 +173,14 @@ describe('hooks/create-hook-app', () => {
       registerAction(constants.SETTINGS, ({ settings }) => {
         expect(settings).toBe(undefined);
       });
-      await runHookApp({ settings: { foo: 1 } });
+      await forrestjs.run({ settings: { foo: 1 } });
     });
 
     it('should handle settings with getters/setters', async () => {
       registerAction(constants.SETTINGS, ({ getConfig, setConfig }) => {
         setConfig('foo', getConfig('foo') + 1);
       });
-      const app = await runHookApp({ settings: { foo: 1 } });
+      const app = await forrestjs.run({ settings: { foo: 1 } });
       expect(app.settings.foo).toBe(2);
     });
 
@@ -141,7 +188,7 @@ describe('hooks/create-hook-app', () => {
       registerAction(constants.SETTINGS, ({ getConfig, setConfig }) => {
         setConfig('new.faa.foo', getConfig('foo') + 1);
       });
-      const app = await runHookApp({ settings: { foo: 1 } });
+      const app = await forrestjs.run({ settings: { foo: 1 } });
       expect(app.settings.new.faa.foo).toBe(2);
     });
   });
@@ -157,9 +204,9 @@ describe('hooks/create-hook-app', () => {
 
     it('should run a required service by reference', async () => {
       const handler = jest.fn();
-      const f1 = ['$S1', handler];
+      const f1 = { hook: '$S1', handler };
 
-      await runHookApp({
+      await forrestjs.run({
         services: [s1],
         features: [f1],
       });
@@ -169,11 +216,11 @@ describe('hooks/create-hook-app', () => {
 
     it('should fail to run a required service by reference', async () => {
       const handler = jest.fn();
-      const f1 = ['$S1', handler];
+      const f1 = { hook: '$S1', handler };
 
       let error = null;
       try {
-        await runHookApp({
+        await forrestjs.run({
           // services: [s1],
           features: [f1],
         });
@@ -186,9 +233,9 @@ describe('hooks/create-hook-app', () => {
 
     it('should ignore an optional service by reference', async () => {
       const handler = jest.fn();
-      const f1 = ['$S1?', handler];
+      const f1 = { hook: '$S1?', handler };
 
-      await runHookApp({
+      await forrestjs.run({
         // services: [s1],
         features: [f1],
       });
@@ -214,10 +261,46 @@ describe('hooks/create-hook-app', () => {
         registerAction('$s1', s1Handler);
       };
 
-      await runHookApp({ services: [s1, s2] });
+      await forrestjs.run({ services: [s1, s2] });
 
       expect(s1Handler.mock.calls.length).toBe(1);
       expect(s2Handler.mock.calls.length).toBe(1);
+    });
+  });
+
+  describe('run declarative features', () => {
+    it('Services and Features should be able to register a single service in a declarative way', async () => {
+      const handler1 = jest.fn();
+
+      await forrestjs.run([
+        {
+          hook: '$INIT_SERVICE',
+          handler: handler1,
+        },
+      ]);
+
+      expect(handler1.mock.calls).toHaveLength(1);
+    });
+
+    it('Services and Features should register multiple hooks in a declarative way', async () => {
+      const handler1 = jest.fn();
+      const handler2 = jest.fn();
+
+      await forrestjs.run([
+        [
+          {
+            hook: '$INIT_SERVICE',
+            handler: handler1,
+          },
+          {
+            hook: '$INIT_SERVICE',
+            handler: handler2,
+          },
+        ],
+      ]);
+
+      expect(handler1.mock.calls).toHaveLength(1);
+      expect(handler2.mock.calls).toHaveLength(1);
     });
   });
 });