+++ packages/kitchensink/app01/feature-home.js
@@ -17,6 +17,7 @@ module.exports = {
   name: 'home',
   hook: '$FASTIFY_GET',
   handler: [
+    { url: '/', handler: async () => 'Home' },
     { url: '/env-var-exists', handler: envVarExists },
     {
       url: '/env-var-not-exists',