+++ packages/kitchensink/service-hash/index.js
@@ -1,10 +1,10 @@
-const { runHookApp } = require('@forrestjs/hooks');
+const forrestjs = require('@forrestjs/hooks');
 const serviceFastify = require('@forrestjs/service-fastify');
 const serviceHash = require('@forrestjs/service-hash');
 
 const featureHome = require('./feature-home');
 
-runHookApp({
+forrestjs({
   trace: 'compact',
   settings: {
     hash: {