+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v4.1.0
+
+- Deprecate array based declarative features [ hook, handler, ... ]
+- Deprecate `runHookApp`
+- Deprecate `createHookApp`
+- Handle a list of declarative hooks as a Feature manifest
+- Expose `forrestjs.run()` API
+
 ## v4.0.0
 
 - Removed babel